refactor: change problem's directory name from domain to entity
