+++ src/main/java/com/kalgory/kp/kalgorykp/converter/ProblemLanguageConverter.java
@@ -1,6 +1,6 @@
 package com.kalgory.kp.kalgorykp.converter;
 
-import com.kalgory.kp.kalgorykp.domain.problem.ProblemLanguage;
+import com.kalgory.kp.kalgorykp.entity.problem.ProblemLanguage;
 
 import javax.persistence.AttributeConverter;
 import java.util.Arrays;
